Move all influx db related code to lib/influx.js
