+++ test/influx.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi } from 'vitest';
-import { reportMetric } from '../lib/metrics.js';
+import { reportMetric, createMetricsFromRequest } from '../lib/influx.js';
 
 describe('reportMetric', () => {
   it('should report metrics correctly', async () => {
@@ -39,5 +39,37 @@ describe('reportMetric', () => {
         })
       })
     );
+  });
+});
+
+describe('createMetricsFromRequest', () => {
+  it('should format request data correctly', () => {
+    const request = {
+      url: 'https://example.com/path?api-key=test-key',
+      method: 'GET',
+      headers: new Map([
+        ['api-key', 'test-key'],
+        ['Authorization', 'Bearer test-token']
+      ])
+    };
+    const response = {
+      status: 200,
+      headers: new Map([
+        ['CF-Cache-Status', 'HIT']
+      ])
+    };
+    const env = {
+      INFLUX_METRIC: 'test_metric'
+    };
+
+    const result = createMetricsFromRequest(request, response, env);
+    expect(result).toContain('test_metric');
+    expect(result).toContain('status_code=200');
+    expect(result).toContain('url=https://example.com/path?api-key\\=test-key');
+    expect(result).toContain('hostname=example.com');
+    expect(result).toContain('pathname="/path"');
+    expect(result).toContain('method=GET');
+    expect(result).toContain('cf_cache_status=HIT');
+    expect(result).toContain('api_key=test-key');
   });
 });
