+++ wrangler.toml.example
@@ -5,7 +5,7 @@ compatibility_flags = ["nodejs_compat"]
 
 [vars]
 REQUEST_URL = "https://example.com"
-INFLUX_METRIC = "example"
+INFLUX_METRIC_NAME = "example"
 INFLUX_URL = "http://localhost:8086"
 INFLUX_TOKEN = "example"
 INFLUX_DATABASE = "example"
